Move test utils of rollback
