+++ components/cdc/src/observer.rs
@@ -334,8 +334,9 @@ mod tests {
     use kvproto::raft_cmdpb::*;
     use std::time::Duration;
     use tikv::storage::kv::TestEngineBuilder;
-    use tikv::storage::mvcc::tests::*;
-    use tikv::storage::txn::tests::{must_commit, must_prewrite_delete, must_prewrite_put};
+    use tikv::storage::txn::tests::{
+        must_commit, must_prewrite_delete, must_prewrite_put, must_rollback,
+    };
 
     #[test]
     fn test_register_and_deregister() {
