+++ src/storage/mvcc/reader/point_getter.rs
@@ -349,7 +349,7 @@ mod tests {
     use crate::storage::mvcc::tests::*;
     use crate::storage::txn::tests::{
         must_acquire_pessimistic_lock, must_commit, must_pessimistic_prewrite_delete,
-        must_prewrite_delete, must_prewrite_lock, must_prewrite_put,
+        must_prewrite_delete, must_prewrite_lock, must_prewrite_put, must_rollback,
     };
 
     fn new_multi_point_getter<E: Engine>(engine: &E, ts: TimeStamp) -> PointGetter<E::Snap> {
