+++ src/storage/mvcc/reader/scanner/forward.rs
@@ -875,7 +875,6 @@ mod latest_kv_tests {
     use super::super::ScannerBuilder;
     use super::*;
     use crate::storage::kv::{Engine, TestEngineBuilder};
-    use crate::storage::mvcc::tests::*;
     use crate::storage::Scanner;
 
     use crate::storage::txn::tests::*;
@@ -1163,8 +1162,9 @@ mod latest_kv_tests {
 mod latest_entry_tests {
     use super::super::ScannerBuilder;
     use super::*;
-    use crate::storage::mvcc::tests::*;
-    use crate::storage::txn::tests::{must_commit, must_prewrite_delete, must_prewrite_put};
+    use crate::storage::txn::tests::{
+        must_commit, must_prewrite_delete, must_prewrite_put, must_rollback,
+    };
     use crate::storage::{Engine, TestEngineBuilder};
     use kvproto::kvrpcpb::Context;
 
@@ -1518,7 +1518,6 @@ mod latest_entry_tests {
 mod delta_entry_tests {
     use super::super::ScannerBuilder;
     use super::*;
-    use crate::storage::mvcc::tests::*;
     use crate::storage::txn::tests::*;
     use crate::storage::{Engine, TestEngineBuilder};
