+++ src/storage/txn/actions/tests.rs
@@ -6,7 +6,7 @@ use super::*;
 use crate::storage::kv::WriteData;
 use crate::storage::mvcc::tests::write;
 use crate::storage::mvcc::{Error, Key, Mutation, MvccTxn, TimeStamp};
-use crate::storage::Engine;
+use crate::storage::{txn, Engine};
 use concurrency_manager::ConcurrencyManager;
 use kvproto::kvrpcpb::Context;
 use pessimistic_prewrite::pessimistic_prewrite;
@@ -460,4 +460,34 @@ pub fn must_pessimistic_prewrite_lock<E: Engine>(
     is_pessimistic_lock: bool,
 ) {
     must_prewrite_lock_impl(engine, key, pk, ts, for_update_ts, is_pessimistic_lock);
+}
+
+pub fn must_rollback<E: Engine>(engine: &E, key: &[u8], start_ts: impl Into<TimeStamp>) {
+    let ctx = Context::default();
+    let snapshot = engine.snapshot(&ctx).unwrap();
+    let start_ts = start_ts.into();
+    let cm = ConcurrencyManager::new(start_ts);
+    let mut txn = MvccTxn::new(snapshot, start_ts, true, cm);
+    txn.collapse_rollback(false);
+    txn::cleanup(&mut txn, Key::from_raw(key), TimeStamp::zero(), false).unwrap();
+    write(engine, &ctx, txn.into_modifies());
+}
+
+pub fn must_rollback_collapsed<E: Engine>(engine: &E, key: &[u8], start_ts: impl Into<TimeStamp>) {
+    let ctx = Context::default();
+    let snapshot = engine.snapshot(&ctx).unwrap();
+    let start_ts = start_ts.into();
+    let cm = ConcurrencyManager::new(start_ts);
+    let mut txn = MvccTxn::new(snapshot, start_ts, true, cm);
+    txn::cleanup(&mut txn, Key::from_raw(key), TimeStamp::zero(), false).unwrap();
+    write(engine, &ctx, txn.into_modifies());
+}
+
+pub fn must_rollback_err<E: Engine>(engine: &E, key: &[u8], start_ts: impl Into<TimeStamp>) {
+    let ctx = Context::default();
+    let snapshot = engine.snapshot(&ctx).unwrap();
+    let start_ts = start_ts.into();
+    let cm = ConcurrencyManager::new(start_ts);
+    let mut txn = MvccTxn::new(snapshot, start_ts, true, cm);
+    assert!(txn::cleanup(&mut txn, Key::from_raw(key), TimeStamp::zero(), false).is_err());
 }
